+++ ios/file_picker/Sources/file_picker/FilePickerPlugin.m
@@ -1,6 +1,7 @@
 #import "FilePickerPlugin.h"
 #import "FileUtils.h"
 #import "ImageUtils.h"
+#import <Flutter/Flutter.h>
 
 #ifdef PICKER_MEDIA
 @import DKImagePickerController;